requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
aiogram==3.13.1
sqlalchemy==2.0.35
alembic==1.13.3
asyncpg==0.29.0
psycopg2-binary==2.9.9
pydantic==2.9.2
pydantic-settings==2.5.2
# Groq совместим с OpenAI SDK — дополнительная библиотека не нужна
openai==1.51.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.12
jinja2==3.1.4
aiohttp==3.10.10
python-dotenv==1.0.1
httpx==0.27.2
